requirements.txt
Django>=5.2
python-dateutil>=2.8.2
django-bootstrap5>=23.3
django-filter>=23.5
